Move LinkAreaConverter to a separate file
